[spinel] update domain prefix spinel property

+++ src/ncp-dummy/DummyNCPInstance.cpp
@@ -129,7 +129,7 @@ DummyNCPInstance::remove_service_on_ncp(uint32_t enterprise_number, const Data &
 }
 
 void
-DummyNCPInstance::add_on_mesh_prefix_on_ncp(const struct in6_addr &addr, uint8_t prefix_len, uint8_t flags,
+DummyNCPInstance::add_on_mesh_prefix_on_ncp(const struct in6_addr &addr, uint8_t prefix_len, uint16_t flags,
 	bool stable, CallbackWithStatus cb)
 {
 	return;
